feat: implement leading slash command parsing functionality
